pyboot/src/dev/engineeringlabs/pyboot/config/pyproject.toml
[project]
name = "pyboot-config"
version = "0.1.0"
description = "Pyboot Config - Configuration management"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
authors = [
    { name = "Software Engineering Labs", email = "dev@engineeringlab.dev" }
]
keywords = ["pyboot", "config", "configuration", "settings", "environment"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Typing :: Typed",
]
dependencies = [
    "pyboot-common",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
toml = ["tomli>=2.0.0"]

[project.urls]
Homepage = "https://github.com/phd-systems/pyboot"
Repository = "https://github.com/phd-systems/pyboot"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
